+++ src/BasicRegisters.php
@@ -3,13 +3,15 @@
 namespace DigipolisGent\Flanders\BasicRegisters;
 
 use DigipolisGent\API\Client\ClientInterface;
-use DigipolisGent\Flanders\BasicRegisters\Filter\Filters;
-use DigipolisGent\Flanders\BasicRegisters\Pager\Pager;
+use DigipolisGent\Flanders\BasicRegisters\Filter\FiltersInterface;
+use DigipolisGent\Flanders\BasicRegisters\Pager\PagerInterface;
 use DigipolisGent\Flanders\BasicRegisters\Request\AddressDetailRequest;
 use DigipolisGent\Flanders\BasicRegisters\Request\AddressListRequest;
+use DigipolisGent\Flanders\BasicRegisters\Request\AddressMatchRequest;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressDetailInterface;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\Addresses;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressMatches;
 
 /**
  * Service to access the Flanders Basic register service.
@@ -35,7 +37,7 @@ final class BasicRegisters implements BasicRegistersInterface
     /**
      * @inheritDoc
      */
-    public function addressList(?Filters $filters = null, ?Pager $pager = null): Addresses
+    public function addressList(?FiltersInterface $filters = null, ?PagerInterface $pager = null): Addresses
     {
         $request = new AddressListRequest($filters, $pager);
         return $this->client->send($request)->addresses();
@@ -48,5 +50,14 @@ final class BasicRegisters implements BasicRegistersInterface
     {
         $request = new AddressDetailRequest($addressId);
         return $this->client->send($request)->addressDetail();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function addressMatch(FiltersInterface $filters = null): AddressMatches
+    {
+        $request = new AddressMatchRequest($filters);
+        return $this->client->send($request)->addressMatches();
     }
 }