+++ src/BasicRegistersFactory.php
@@ -5,6 +5,7 @@ namespace DigipolisGent\Flanders\BasicRegisters;
 use DigipolisGent\API\Client\ClientInterface;
 use DigipolisGent\Flanders\BasicRegisters\Handler\AddressDetailHandler;
 use DigipolisGent\Flanders\BasicRegisters\Handler\AddressListHandler;
+use DigipolisGent\Flanders\BasicRegisters\Handler\AddressMatchHandler;
 
 /**
  * Factory to get the BasicRegisters.
@@ -20,6 +21,7 @@ final class BasicRegistersFactory
     {
         $client->addHandler(new AddressListHandler());
         $client->addHandler(new AddressDetailHandler());
+        $client->addHandler(new AddressMatchHandler());
 
         return new BasicRegisters($client);
     }