+++ src/BasicRegistersInterface.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters;
 
-use DigipolisGent\Flanders\BasicRegisters\Filter\Filters;
-use DigipolisGent\Flanders\BasicRegisters\Pager\Pager;
+use DigipolisGent\Flanders\BasicRegisters\Filter\FiltersInterface;
+use DigipolisGent\Flanders\BasicRegisters\Pager\PagerInterface;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressDetailInterface;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\Addresses;
 use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressMatches;
 
 /**
  * Service to access the Flanders Basic register service.
@@ -18,15 +19,15 @@ interface BasicRegistersInterface
     /**
      * Get a list of addresses
      *
-     * @param \DigipolisGent\Flanders\BasicRegisters\Filter\Filters|null $filters
+     * @param \DigipolisGent\Flanders\BasicRegisters\Filter\FiltersInterface|null $filters
      *   Optional filters to limit the returned addresses by.
-     * @param \DigipolisGent\Flanders\BasicRegisters\Pager\Pager|null $pager
+     * @param \DigipolisGent\Flanders\BasicRegisters\Pager\PagerInterface|null $pager
      *   Optional pager.
      *
      * @return \DigipolisGent\Flanders\BasicRegisters\Value\Address\Addresses
      *   Collection of found addresses.
      */
-    public function addressList(?Filters $filters = null, ?Pager $pager = null): Addresses;
+    public function addressList(?FiltersInterface $filters = null, ?PagerInterface $pager = null): Addresses;
 
     /**
      * Get the address details by a given AddressId value.
@@ -38,4 +39,20 @@ interface BasicRegistersInterface
      *   The address detail value.
      */
     public function addressDetail(AddressId $addressId): AddressDetailInterface;
+
+    /**
+     * Look up addresses that match (partial) filter values.
+     *
+     * NOTE: The matcher requires that minimal following filters are provided:
+     *   (LocalityName AND/OR LocalityNisCode AND/OR PostalCode)
+     *   AND
+     *   (StreetName AND/OR StreetPatrimonyCode AND/OR StreetNationalRegisterCode)
+     *
+     * @param \DigipolisGent\Flanders\BasicRegisters\Filter\FiltersInterface $filters
+     *   Optional filters to limit the returned addresses by.
+     *
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressMatches
+     *   The matching addresses.
+     */
+    public function addressMatch(FiltersInterface $filters = null): AddressMatches;
 }