+++ tests/Filter/StreetNationalRegisterCodeFilterTest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Filter;
+
+use DigipolisGent\Flanders\BasicRegisters\Filter\StreetNationalRegisterCodeFilter;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Filter\StreetNationalRegisterCodeFilter
+ */
+class StreetNationalRegisterCodeFilterTest extends TestCase
+{
+    /**
+     * Filter has proper argument name.
+     *
+     * @test
+     */
+    public function hasProperArgumentName(): void
+    {
+        $filter = new StreetNationalRegisterCodeFilter('FooBar');
+
+        $this->assertEquals('rrStraatcode', $filter->name());
+    }
+}