+++ CHANGELOG.md
@@ -36,5 +36,6 @@ All Notable changes to `digipolisgent/flanders-basicregisters` package.
 * Added the configuration.
 * Added service method to get a list of Addresses.
 * Added service method to get the details of a single address.
+* Added service method to get (partial) addresses that match filter(s).
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/master...develop